support import trojan uri,share trojan support multi outbounds restore ping
